stayhome/requirements.txt
asgiref==3.2.5
astroid==2.3.3
Babel==2.8.0
Django==3.0.4
django-js-asset==1.2.2
django-mptt==0.11.0
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-recaptcha==2.0.6
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
phonenumbers==8.11.5
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
gunicorn==20.0.4
whitenoise==5.0.1
sentry-sdk==0.14.2
python-memcached==1.59
django-modeltranslation==0.14.4